requirements.txt
streamlit
yfinance
plotly
python-dotenv
feedparser
google-generativeai
newspaper3k